Limit memory for worker processes by default
Our workers for compilation, testing etc. had no default
memory settings until now, which mean that we would use
the default given by the JVM, which is 1/4th of system memory
in most cases. This is way too much when running with a high
number of workers. Limiting this to 256m should be enough for
small to medium sized projects. Larger projects will have to
tweak this value.